server/requirements.txt
fastapi
uvicorn
boto3
pydantic-settings
python-dotenv
# db
sqlalchemy
psycopg2-binary
redis